# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "BCR API",
      "version": "1.0",
      "contact": {
        "name": "Sena Putra",
        "url": "https://gitlab.com/senaaputra",
        "email": "senaaputra@gmail.com"
      }
    },
...
    "consumes": [
      "application/json",
      "multipart/form-data"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
...
			"/api/user/logout": {
				"delete": {
					"summary": "Logout",
					"description": "Remove existing token by id",
					"operationId": "deleteProduct",
					"responses": {
						"200": {
							"description": "A successful response.",
							"schema": {
								"properties": {}
							}
						},
						"default": {
							"description": "An unexpected error response.",
							"schema": {
								"$ref": "#/definitions/rpcStatus"
							}
						}
					},
					"tags": ["Auth"]
				}
			},
...
			"/api/product/create": {
				"post": {
					"summary": "Create Product",
					"description": "Add new car on the database",
					"operationId": "createProduct",
					"security": [{ "BearerTokenAuth": [] }],
					"responses": {
						"200": {
							"description": "A successful response.",
							"schema": {
								"$ref": "#/definitions/coreProduct"
							}
						},
						"default": {
							"description": "An unexpected error response.",
							"schema": {
								"$ref": "#/definitions/rpcStatus"
							}
						}
					},
					"parameters": [
						{
							"name": "body",
							"in": "body",
							"required": true,
							"schema": {
								"$ref": "#/definitions/coreCreateProductRequest"
							}
						}
					],
					"tags": ["Products"]
				}
			},
...
			"/api/product/delete/{id}": {
				"delete": {
					"summary": "Delete Product",
					"description": "Remove existing product by id",
					"operationId": "deleteProduct",
					"security": [{ "BearerTokenAuth": [] }],
					"responses": {
						"200": {
							"description": "A successful response.",
							"schema": {
								"properties": {}
							}
						},
						"default": {
							"description": "An unexpected error response.",
							"schema": {
								"$ref": "#/definitions/rpcStatus"
							}
						}
					},
					"parameters": [
						{
							"name": "id",
							"in": "path",
							"required": true,
							"type": "integer",
							"format": "int64"
						}
					],
					"tags": ["Products"]
				}
			},
			"/api/product/{id}": {
				"put": {
					"summary": "Update Product",
					"description": "Change product information",
					"operationId": "updateProduct",
					"security": [{ "BearerTokenAuth": [] }],
					"responses": {
						"200": {
							"description": "A successful response.",
							"schema": {
								"$ref": "#/definitions/coreProduct"
							}
						},
						"default": {
							"description": "An unexpected error response.",
							"schema": {
								"$ref": "#/definitions/rpcStatus"
							}
						}
					},
					"parameters": [
						{
							"name": "id",
							"in": "path",
							"required": true,
							"type": "integer",
							"format": "int64"
						},
						{
							"name": "body",
							"in": "body",
							"required": true,
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string"
									},
									"price": {
										"type": "integer",
										"format": "int64"
									},
									"description": {
										"type": "string"
									},
									"categoryId": {
										"type": "integer",
										"format": "int64"
									},
									"image": {
										"type": "string"
									}
								}
							}
						}
					],
					"tags": ["Products"]
				}
			},
...
		},
    "definitions": {
...
      "coreProduct": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "format": "int64"
          },
          "description": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "category": {
            "$ref": "#/definitions/coreCategory"
          },
          "user": {
            "$ref": "#/definitions/coreWhoAmIResponse"
          }
        }
      },
      "coreCategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "coreWhoAmIResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "protobufAny": {
        "type": "object",
        "properties": {
          "@type": {
            "type": "string"
          }
        },
        "additionalProperties": {}
      },
      "rpcStatus": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/protobufAny"
            }
          }
        }
      }
    },
    "securityDefinitions": {
      "BearerTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    }
  }